// src/models/tvshowDetails.models.ts
export interface TvShowDetails {
    backdropPath: string;
    episodeRunTime: number[];
    genres: string[];
    name: string;
    numberOfEpisodes: number;
    numberOfSeasons: number;
    originalLanguage: string;
    overview: string;
    posterPath: string;
    status: string;
    tagline: string;
    type: string;
    trailersURL: string[];
}